Baked another cake
Fixed the issue of the cake not appearing on the screen by reseting
Spritebuilder. Also (should have) fixed Jenny’s issue of it not
building on her laptop but now there’s a bug with the levels and it not
being able to get the properties.
—> Temporarily deleted the Level class from Level1 for now for testing
purposes. Will try to debug this tomorrow.